Split libs into compile-time and runtime packages
