+++ lib/array/runtime/runtime.go
@@ -1,7 +1,7 @@
 package array
 
 import (
-	"ballerina-lang-go/runtime/api"
+	"ballerina-lang-go/runtime"
 	"fmt"
 )
 
@@ -10,15 +10,15 @@ const (
 	moduleName = "lang.array"
 )
 
-func initArrayModule(rt *api.Runtime) {
-	api.RegisterExternFunction(rt.Registry, orgName, moduleName, "push", func(args []any) (any, error) {
+func initArrayModule(rt *runtime.Runtime) {
+	runtime.RegisterExternFunction(rt, orgName, moduleName, "push", func(args []any) (any, error) {
 		if arr, ok := args[0].(*[]any); ok {
 			*arr = append(*arr, args[1:]...)
 			return nil, nil
 		}
 		return nil, fmt.Errorf("first argument must be an array")
 	})
-	api.RegisterExternFunction(rt.Registry, orgName, moduleName, "length", func(args []any) (any, error) {
+	runtime.RegisterExternFunction(rt, orgName, moduleName, "length", func(args []any) (any, error) {
 		if arr, ok := args[0].(*[]any); ok {
 			return len(*arr), nil
 		}
@@ -27,5 +27,5 @@ func initArrayModule(rt *api.Runtime) {
 }
 
 func init() {
-	api.RegisterModuleInitializer(initArrayModule)
+	runtime.RegisterModuleInitializer(initArrayModule)
 }
